x402-data-provider/gwdc_tina/requirements.txt
openai>=1.0.0
yfinance>=0.2.30
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn
feedparser>=6.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
